+++ lib/src/widgets/responsive_constrained_box.dart
@@ -6,7 +6,7 @@ import 'package:flutter_gehry/flutter_gehry.dart';
 /// If no [BoxConstraints] has been provided for the current breakpoint, the child will be directly returned, without any [ConstrainedBox].
 /// If [child] is also null, a [SizedBox.shrink] will be returned.
 class ResponsiveConstrainedBox extends StatelessWidget with ResponsiveStatelessWidgetMixin {
-  final ResponsiveBreakpointsData<BoxConstraints> data;
+  final BreakpointsData<BoxConstraints> data;
 
   /// Child of the [ConstrainedBox].
   final Widget? child;
@@ -19,7 +19,7 @@ class ResponsiveConstrainedBox extends StatelessWidget with ResponsiveStatelessW
     BoxConstraints? lg,
     BoxConstraints? xl,
     required this.child,
-  })  : data = ResponsiveBreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
+  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
         super(key: key);
 
   @override