+++ lib/src/widgets/responsive_widget.dart
@@ -4,7 +4,7 @@ import 'package:flutter_gehry/flutter_gehry.dart';
 /// Displays the widget for the current breakpoint,
 /// or a [SizedBox.shrink)] if no data has been provided for the breakpoint.
 class ResponsiveWidget extends StatelessWidget with ResponsiveStatelessWidgetMixin {
-  final ResponsiveBreakpointsData<Widget> data;
+  final BreakpointsData<Widget> data;
 
   ResponsiveWidget({
     Key? key,
@@ -13,7 +13,7 @@ class ResponsiveWidget extends StatelessWidget with ResponsiveStatelessWidgetMix
     Widget? md,
     Widget? lg,
     Widget? xl,
-  })  : data = ResponsiveBreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
+  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
         super(
           key: key,
         );